backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-multipart>=0.0.18
markitdown[all]>=0.1.0
llama-cpp-python>=0.3.0
openai>=1.60.0
pydantic-settings>=2.7.0
pytesseract>=0.3.10
Pillow>=10.0.0
